// public/app/features/baseline/types.ts
export interface BaselineEntryFields {
  startDate: string;
  endDate: string;
  kwh: string;
  minKw: string;
  maxKw: string;
  avgKw: string;
  avgKva: string;
  pf: string;
  minPf: string;
  maxPf: string;
  rate: string;
  energyRate: string;
  fuelRate: string;
  ippRate: string;
  ippVariableRate: string;
  ippVariableCharge: string;
  energyCharge: string;
  currentCharges: string;
}